pyproject.toml
[project]
name = "voice-txt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "numpy>=2.4.2",
    "openai>=2.24.0",
    "openai-whisper>=20250625",
    "pyaudio>=0.2.14",
    "pyautogui>=0.9.54",
    "pyinstaller>=6.19.0",
    "pynput>=1.8.1",
    "pyperclip>=1.11.0",
    "pystray>=0.19.5",
    "python-dotenv>=1.2.2",
    "setuptools-rust>=1.12.0",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
    "customtkinter>=5.2.2",
    "pillow>=10.2.0",
]
